pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "baobab-barcode"
version = "0.1.0"
description = "Bibliothèque Python pour la génération et la lecture de codes-barres."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Baobab Barcode contributors" }]
keywords = ["barcode", "code128", "qrcode"]

dependencies = [
    "pillow>=10.0.0,<12.0.0",
    "pyzbar>=0.1.9,<0.2.0",
    "python-barcode>=0.15.1,<0.17.0",
    "qrcode>=7.4.2,<9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<9.0.0",
    "pytest-cov>=4.0.0,<7.0.0",
    "black>=24.0.0,<26.0.0",
    "pylint>=3.0.0,<5.0.0",
    "mypy>=1.8.0,<2.0.0",
    "bandit>=1.7.0,<2.0.0",
    "flake8>=7.0.0,<8.0.0",
    "Flake8-pyproject>=1.2.3,<2.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
baobab_barcode = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
addopts = [
    "--cov=baobab_barcode",
    "--cov-report=term-missing",
    "--cov-fail-under=90",
]

[tool.coverage.run]
branch = true
source = ["src/baobab_barcode"]

[tool.coverage.report]
fail_under = 90
show_missing = true
skip_empty = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
per-file-ignores = ["*/__init__.py:F401"]

[tool.mypy]
python_version = "3.11"
strict = true
explicit_package_bases = true
mypy_path = "src"
files = ["src/baobab_barcode", "tests"]

[[tool.mypy.overrides]]
module = "barcode.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "qrcode.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyzbar.*"
ignore_missing_imports = true

[tool.pylint.main]
fail-under = 9.0
jobs = 0
py-version = "3.11"
source-roots = ["src"]

[tool.pylint.format]
max-line-length = 100

[tool.pylint."messages control"]
disable = [
    "too-few-public-methods",
]

[tool.bandit]
exclude_dirs = ["tests"]
